requirements.txt
openai>=1.2.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
requests>=2.31.0
asyncio>=3.0.0
streamlit>=1.30.0
